Crate libipld

The Ipld crate.

Re-exports

pub use block::Block;

pub use codec_impl::IpldCodec;

pub use path::DagPath;

pub use path::Path;

pub use store::DefaultParams;

pub use libipld_cbor as cbor;

pub use libipld_json as json;

Modules

block

Block validation

cid

cid

codec

Ipld codecs.

codec_impl

IPLD Codecs.

convert

Conversion to and from ipld.

error

Ipld error definitions.

ipld

Ipld representation.

link

Typed cid.

path

Path

prelude

Prelude

raw

Implements the raw codec.

raw_value

misc stuff

serde

Serde (de)serialization for crate::ipld::Ipld.

store

Store traits.

Macros

ipld

Structs

Link

Typed cid.

Enums

Ipld

Ipld

Type Aliases

Cid

A Cid that contains a multihash with an allocated size of 512 bits.

Multihash

A Multihash with the same allocated size as the Multihashes produces by this derive.

Result

Result<T, Error>

Derive Macros

DagCbor

Multihash