Trait libipld::prelude::Decode

pub trait Decode<C>: Sizedwhere
    C: Codec,{
    // Required method
    fn decode<R>(c: C, r: &mut R) -> Result<Self, Error>
       where R: Read + Seek;
}

Decode trait.

This trait is generic over a codec, so that different codecs can be implemented for the same type.

Required Methods

fn decode<R>(c: C, r: &mut R) -> Result<Self, Error>where R: Read + Seek,

Decode from an impl Read.

It takes a specific codec as parameter, so that the Decode can be generic over an enum that contains multiple codecs.