Trait libipld::prelude::Encode

pub trait Encode<C>where
    C: Codec,{
    // Required method
    fn encode<W>(&self, c: C, w: &mut W) -> Result<(), Error>
       where W: Write;
}

Encode trait.

This trait is generic over a codec, so that different codecs can be implemented for the same type.

Required Methods

fn encode<W>(&self, c: C, w: &mut W) -> Result<(), Error>where W: Write,

Encodes into a impl Write.

It takes a specific codec as parameter, so that the Encode can be generic over an enum that contains multiple codecs.