Trait libipld::raw_value::SkipOne

pub trait SkipOne: Codec {
    // Required method
    fn skip<R>(&self, r: &mut R) -> Result<(), Error>
       where R: Read + Seek;
}

trait to implement to skip a single item at the current position

Required Methods

fn skip<R>(&self, r: &mut R) -> Result<(), Error>where R: Read + Seek,

assuming r is at the start of an item, advance r to the end