Type Alias libipld::Cid

pub type Cid = Cid<64>;

A Cid that contains a multihash with an allocated size of 512 bits.

This is the same digest size the default multihash code table has.

If you need a CID that is generic over its digest size, use CidGeneric instead.

Aliased Type

struct Cid { /* private fields */ }

Implementations

impl<const S: usize> Cid<S>

pub const fn new_v0(hash: Multihash<S>) -> Result<Cid<S>, Error>

Create a new CIDv0.

pub const fn new_v1(codec: u64, hash: Multihash<S>) -> Cid<S>

Create a new CIDv1.

pub const fn new( version: Version, codec: u64, hash: Multihash<S> ) -> Result<Cid<S>, Error>

Create a new CID.

pub fn into_v1(self) -> Result<Cid<S>, Error>

Convert a CIDv0 to a CIDv1. Returns unchanged if already a CIDv1.

pub const fn version(&self) -> Version

Returns the cid version.

pub const fn codec(&self) -> u64

Returns the cid codec.

pub const fn hash(&self) -> &Multihash<S>

Returns the cid multihash.

pub fn read_bytes<R>(r: R) -> Result<Cid<S>, Error>where R: Read,

Reads the bytes from a byte stream.

pub fn write_bytes<W>(&self, w: W) -> Result<usize, Error>where W: Write,

Writes the bytes to a byte stream, returns the number of bytes written.

pub fn encoded_len(&self) -> usize

Returns the length in bytes needed to encode this cid into bytes.

pub fn to_bytes(&self) -> Vec<u8, Global>

Returns the encoded bytes of the Cid.

pub fn to_string_of_base(&self, base: Base) -> Result<String, Error>

Convert CID into a multibase encoded string

Example

use cid::Cid;
use multibase::Base;
use multihash::{Code, MultihashDigest};

const RAW: u64 = 0x55;

let cid = Cid::new_v1(RAW, Code::Sha2_256.digest(b"foo"));
let encoded = cid.to_string_of_base(Base::Base64).unwrap();
assert_eq!(encoded, "mAVUSICwmtGto/8aP+ZtFPB0wQTQTQi1wZIO/oPmKXohiZueu");